Fix checking for undefined variables
Since we now create a global variable reference if a variable does not
exist, we can't just check for the free variables of the top-level
function. Instead, we keep track of all referenced global variables, and
at the end make sure all are in "params" of the top-level function (all
defined variables are added to "params").