Fix let variable
We had broken the let variable because it was no longer considered a
free variable, but a global variable, and as such not captured in the
environment.

We now add a dummy function as parent to the new let function while
compiling its body. The dummy function has the let variable as its sole
parameter, and so we capture the let variable as a free variable.