Prepare for adding "apply"
This lays some foundations for adding "apply". First of all, instead of
using va_args, etc, we have a new set of custom macros. Also, a
call_flags parameter has been added to all functions. This can later be
used to pass an array of arguments as a single argument to a function,
and use that in place of va_list.