Add gensym and uninterned symbols
We now have uninterned symbols; symbols that are actually objects and
are not registered in the symbol table. These can be differentiated from
normal symbols by `uninterned-symbol?`

`(gensym)` returns a fresh uninterned symbol with a name consisting of a
prefix "g" and a counter. `(gensym "foo")` returns a fresh uninterned
symbol named "foo". Notice that an uninterned symbol is never `eq?` to
another symbol even if it has the same name.

Uninterned symbols are printed with a #: prefix.