Fix reading sharp identifiers (#t and #f)
We were reading the #, the peeking the t/f, and then we would push back
the t/f. We now read the #, read the t/f, and then push them back in the
correct order.

Notice that this is not guaranteed to be supported by all platforms. C99
has this to say about ungetc:

    One character of pushback is guaranteed. If the ungetc function is
    called too many times on the same stream without an intervening read
    or file positioning operation on that stream, the operation may
    fail.