Add wrapped values
A wrapped value is a value which is paired with another value as its
"kind" or tag. This is basically a pair, with a distinct type and is
intended to be used for creating new user defined types (like records).
There are four primitives:

(wrap obj kind)  ;; wrap obj with the given "kind"
(unwrap w)       ;; extract the wrapped object
(wrapped-kind w) ;; extract the wrapped "kind"
(wrapped? w)     ;; returns true if w is a wrapped object

"kind" can be any object, but a particularly suitable object for it is a
symbol. If you use a unique uninterned symbol, you'll get a completly
unique type.